Code cleanup, comments/documentation and licensing
- Cleaned up code (consistent naming, etc)
- Commented/documented code everywhere
- Added licensing information to all source files

+++ LightsController.h
@@ -1,4 +1,18 @@
 // LightsController.h
+// Copyright (C) 2015  Alex Goris
+// This file is part of FlyballETS-Software
+// FlyballETS-Software is free software : you can redistribute it and / or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.If not, see <http://www.gnu.org/licenses/>
 
 #ifndef _LIGHTSCONTROLLER_h
 #define _LIGHTSCONTROLLER_h
@@ -14,7 +28,7 @@ class LightsControllerClass
 protected:
 
 public:
-	void init(int iLatchPin, int iClockPin, int iDataPin);
+	void init(uint8_t iLatchPin, uint8_t iClockPin, uint8_t iDataPin);
 
    //Overal state of this class
    enum OverallStates{
@@ -49,11 +63,11 @@ public:
    
 private:
    //Pin connected to ST_CP of 74HC595
-   int _iLatchPin = 12;
+   uint8_t _iLatchPin = 12;
    //Pin connected to SH_CP of 74HC595
-   int _iClockPin = 13;
+   uint8_t _iClockPin = 13;
    //Pin connected to DS of 74HC595
-   int _iDataPin = 11;
+   uint8_t _iDataPin = 11;
 
    //This byte contains the combined states of all ligths at any given time
    byte _byCurrentLightsState = 256;
